use positive and negative flags for player groups
Because it turns out we don't need to store any other flags but the
player group (ko is stored in the state array now), we can simply use
positive flags to denote player1 (black) and negative groups to denote
player2.  This will be much more convenient to work with on a device,
than the even/odd encoding we had before.